test: Update test database connection properties
This commit updates the Postgres database connection properties in the test environment configuration. This includes using environment variables for all configurable properties and preparing for the usage of SSL. Additionally, the scope of entities and migrations files has been broadened, and only error-level logs will be outputted during the test run.