chore: Update relationships to User in goal.entity to be eager-loaded
This commit modifies the @ManyToOne and @ManyToMany relationships from goal.entity to User in the backend codebase to be eager-loaded. This means the related User entities will automatically be loaded whenever a goal is retrieved from the database, which can significantly improve performance in scenarios where the User data are frequently accessed.