feat(frontend): developed a left-hand panel
In this commit, the unused React logo SVG was removed from the repository. Changes have been made to the homepage layout, introducing a new left hand panel and modifying the header and other css styling for an enhanced look and feel. Additionally, unused code in 'app.tsx' was removed, and relevant routing logic was revised.