Add user role modification to UserController
Added a new method 'modifyRoles' to the UserController that allows user roles to be updated. Set utilization of the Admin role for this action which enhances the system's role-based access control (RBAC). Introduced a new UpdateUserRoleDto to manage role update requests. Created a new UserService's method to segregate the concerns