feat: add subgoal cards to home component
This commit introduces a new SubgoalCard component to display individual subgoals on the home page. All required files were updated, including the Goal service to fetch subgoal data, and the HomeComponent to render the list of subgoals. The MatCardModule was also added into MaterialModule to accommodate the new card component.