feat: beautify subgoal cards UI and enhance details
The updates include implementing a Grid List to display the SubgoalCard components and introducing new CSS rules to improve the look and feel. Simultaneously, adjustments have been made in the Goal service to ensure the compatibility of lengthier subgoal details with the enhanced UI. Import of MatGridListModule was also added to the MaterialModule.