Refactored app structure and added login component
In this commit, the project structure was reorganized moving some components into a "shared" directory. Furthermore, HttpClientModule was added to the project configuration and MaterialModule was moved to the shared directory. A new login component and related spec files were also added. Finally, in package.json the start script was renamed to start:dev.